requirements.txt
discord.py>=2.5.0
pymongo>=4.12.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
aiohttp>=3.9.0
itsdangerous>=2.2.0

# ML training (only needed on the machine where you train)
scikit-learn>=1.5.0
numpy>=1.26.0
joblib>=1.4.0
sentence-transformers>=2.2.0
